# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: QA Dashboard API
  version: 3.0.0
paths:
...
  /api/dashboard/multi:
    get:
      tags:
      - dashboard
      summary: Multi Project Dashboard
      operationId: multi_project_dashboard_api_dashboard_multi_get
      parameters:
      - name: project_ids
        in: query
        required: false
        schema:
          type: array
          items:
            type: integer
          default: []
          title: Project Ids
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/dashboard/compare:
    get:
      tags:
      - dashboard
      summary: Compare Dashboard
      operationId: compare_dashboard_api_dashboard_compare_get
      parameters:
      - name: project_ids
        in: query
        required: false
        schema:
          type: array
          items:
            type: integer
          default: []
          title: Project Ids
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/dashboard/{project_id}:
    get:
      tags:
      - dashboard
      summary: Get Dashboard
      operationId: get_dashboard_api_dashboard__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/dashboard/{project_id}/quality-rates:
    get:
      tags:
      - dashboard
      summary: Get Quality Rates
      operationId: get_quality_rates_api_dashboard__project_id__quality_rates_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      - name: preprod_milestones
        in: query
        required: false
        schema:
          type: array
          items:
            type: integer
          default: []
          title: Preprod Milestones
      - name: prod_milestones
        in: query
        required: false
        schema:
          type: array
          items:
            type: integer
          default: []
          title: Prod Milestones
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/dashboard/{project_id}/annual-trends:
    get:
      tags:
      - dashboard
      summary: Get Annual Trends
      operationId: get_annual_trends_api_dashboard__project_id__annual_trends_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/dashboard/{project_id}/trends:
    get:
      tags:
      - dashboard
      summary: Get Trends
      operationId: get_trends_api_dashboard__project_id__trends_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      - name: granularity
        in: query
        required: false
        schema:
          type: string
          default: daily
          title: Granularity
      - name: from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From
      - name: to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/dashboard/{project_id}/stream:
    get:
      tags:
      - dashboard
      summary: Stream Dashboard
      operationId: stream_dashboard_api_dashboard__project_id__stream_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/anomalies/{project_id}:
    get:
      tags:
      - anomalies
      summary: Get Anomalies
      operationId: get_anomalies_api_anomalies__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer